merge scalahost and scalahostNsc
Some time ago, we split scalahost into scalahostNsc (the compiler plugin
that produces semantic databases) and scalahost (the library that reads
semantic databases): https://github.com/scalameta/scalameta/pull/775.

This split was motivated by infrastructural reasons (scalahostNsc must
be a fat jar to work well with sbt, but scalahost must be a normal jar
to avoid issues like https://github.com/scalacenter/scalafix/issues/100),
but in hindsight it was a great idea because of a completely different reason.

As we’ve learned recently, the subsystem that produces semantic dbs
and the subsystem that consumes semantic dbs can and should be separated.
As a result, we moved the consumer part to the `semantic` module of scala.meta,
effectively leaving scalahost without its raison d'être.

This commit recognizes this situation and removes scalahost altogether,
leaving us with just scalahostNsc.