Enable js-bundler in both io and scalameta modules.
The tests run fine in ioJS but scalametaJS/test triggers the error
```
[error] Error: Cannot find module 'shelljs'
```
I suspect this is because npmDependencies and scalajs-bundler is only
enabled in the io project. This commit works around that error by
enabling scalajs-bundler in io and scalameta. Ideally, we should only
have to enable it once in the io project, I will ask Julien at the
SC who implemented scalajs-bundler if we can find a better workaround.