collapse scala.meta.internal.semantic.mirrors and the like
Now that we've removed auxiliary stuff in Mirror, the Mirror
implementation itself is becoming increasingly trivial.

As a result, I got rid of dedicated packages that hosted helper
classes in both semantic and scalahost.

This set off a chain reaction of reevaluating other packages in scalahost
and making things even simpler.