split Database into vfs/schema/meta layers
This commit reformulates existing serialization/deserialization logic
of semantic databases in a more robust way, paving the way to making
semantic databases portable and potentially fat. Take a look at the
new ScalahostPipeline. This is a thing of beauty!

In order to achieve this, we needed a major rethink of what is the essence
a semantic database, how this all is related to mirrors, and other fundamental
questions. Needless to say, this led to more refactorings and cleanups
that affected the entire scala.meta.semantic.

Moreover, scala.meta.io gained some new members, including Classpath,
Sourcepath and Multipath. This should make it easier to correctly handle
jars and other pesky aspects of dealing with the JVM infrastructure.