Improve Xml parser performances
I believe this can improve the performances of the xml parser.
We instantiate a new parser for every single xml literals. I believe
most of the time is spent during initialization since our xml parsers are
short lived and not reusable. These changes reduce initialization time
and memory consumption as it avoids pre-computing the lookup tables.

A better solution would be to have a reusable parser. But we would need
to remove the parser dependency on `Dialect`.