relax overly strong typing in pattern signatures
1) Back then, Denys was shaming me a lot for introducing Pat.Type,
and he was right - that's an abomination that doesn't have place in meta.

In this commit, I removed Pat.Type and introduced something called Type.Var
which is a subtype of Type and can be used instead of Pat.Var.Type.
Consequently, Pat.Var.Term got renamed to just Pat.Var.

Admittedly, user experience around variables in type patterns got worse,
because we no longer have a quasiquote-based way of constructing
these trees. However, this is such a fringe language feature,
and the simplification of the tree API is so massive, that I'm willing
to compromise on this.

Safety by construction isn't compromised, because I've added a parent check
to the newly introduced Type.Var.

2) Pat.Bind.lhs, Decl.Var/Var.pats got their types dumbed down from Pat.Var
to just Pat in order to simplify the API surface that we expose in
quasiquotes. Safety by construction is again not compromised.

3) Pat.Extract now only has two fields: fun and args - to be similar
to Term.Apply. The previous ref and targs fields got folded into fun.
Safety by construction remains intact.