remove Name.Qualifier
This commit further continues the trend of simplifying our public API
as seen in quasiquotes.md.

This time we give up on unification of private[this] and private[within]
(which was arguable to begin with) and as result get rid of Name.Qualifier.
Honestly, without looking into docs, I wasn't even sure where this trait
is used and who implements it.

Static safety by construction is a little bit hurt by this change,
because now we can write weird stuff like mod"private[${ctor"foo"}]".
However, it is also a little bit improved, because now we can't write other
weird stuff like mod"${q"this"}.this".