remove Name.Indeterminate from the public API
Newcomers are frequently confused by the wide array of names that
scalameta uses to represent different parts of the language.

This commit aims to address the confusion by replacing all occurences
of the scary Name.Indeterminate with just Name. Indeterminate's close friend,
Name.Anonymous was left behind the scenes by one of the previous commits.

This way we can explain to our users that there are just names, term names,
type names and ctor names. Most of the time, one operates with either
Term.Name or Type.Name. In the rare cases when one needs to go beyond
that, they don't need to know the specifics - just having a Name
that doesn't fit anywhere will suffice.

Ctor names don't quite fit into this picture at the moment, but I'm planning
to refactor ctors anyway, so I'll take care of that in one of the follow-up
commits.