remove xml-related RangePosition
I was very surprised to see something called RangePosition in our
xml parser. We have our own Position.Range, so ideological duplication
is going to be suboptimal.

Since this class was used only twice, I replaced it with a tuple
and removed the original definition. This saved some mental cycles
necessary for coming up with an appropriate name.