merge Mod.XXXThis/XXXWithin into Mod.XXX
No idea why I decided to split these modifiers a couple days ago.
The only thing I remember is that it was close to 12am, and I was hacking
almost fourteen hours straight.

This exposes a new field type to quasiquotes, but I don't see a problem
with that. This type (scala.meta.Ref) doesn't fit anywhere else, which
means that it's impossible to misuse it.