inline Point into Position
Back then, when I introduced Point as a separate entity, I was mainly
thinking how to avoid having to write `startLine`, `startColumn`
and then `endLine`, `endColumn`.

Indeed, we ended up with a more minimalistic API for Position,
but the quality of the overall API suffered significantly:
  * The most frequent operations, i.e. `Point.start` and `Point.end`,
    have actually become more complicated, because one would need to say
    `Point.start.offset` and `Point.end.offset`.
  * Even obtaining lines and columns became harder, because one would say
    `Point.start.line` instead of `Point.startLine`.

The only beneficial thing that I can see is the ability to treat points
independently from the enclosing position, e.g. Point.formatMessage.
However, I don't think it overweighs the downsides.

It's time to put an end to this.