Setup Travis CI.
We are experiencing resolution errors with the platform-ci, which is
slowing us down. My hypothesis is that our bandwith is being limited,
partly due to bad caching setup. The dotty-ci is apparently downloading
50gb a day. Instead of spending more days on the drone CI, I'm setting
up travis again. I disable the ci-slow tests since those are too heavy
for the free-tier Travis servers.