replace Seq with List
On the day one, we decided to use scala.collection.immutable.Seq
instead of the well-known, no-brainer scala.List.

Why did we decide to complicate things? Well, we did that in order to
leave our options open with respect to the exact representation
of various collections in our API.

This seems overly general, but the main target of this design decision
was very specific. Back then, we planned to persist typed ASTs for all programs
(something that has become TASTY about a year later). We also planned to
unify symbols and trees, which meant that introspection operations like
Tree.symbol or Type.members would return trees. As a result, we really needed
lazy loading for things like Template.stats.

Three years later, we live in a completely different world.
We no longer persist ASTs (that's the responsibility for TASTY),
and we no longer unify symbols and trees.

However, we still pay the complexity tax associated with the flexibility
that's no longer necessary. For instance, check out the recently submitted
https://github.com/scalameta/tutorial/issues/18 - it's really impressive
that such a small thing can lead to such confusion.

As a result, during the 2.x cleanup, I decided to replace all usages
of Seq in our public API with the good old List.

Legitimate occurrences of "seq" remain untouched, namely:
  * Show.Sequence
  * *Seq methods in ScalametaParser (they follow the terminology in SLS)
  * Pat.SeqWildcard (also follows the terminology of SLS)
  * Base traits of the Tokens collection (we can't use lists there,
    because we need O(1) performance of indexing)