Document SugarCodePrinter changes.
It may have been a bad idea to reformat the sources. I was struggling to
follow some of the logic so I took the liberty to run scalafmt on it
with `rewrite.rules = [AvoidInfix]` enabled. The eventual modifications
are so small we can if we want to apply them on the original source.