fix #583 for annotations
It turns out that the quasiquote engine was inferring an incorrect type
for `annot` in `case mod"@$annot"`.

The type of `annot` should be Mod.Annot, but it was incorrectly inferred
as Tree. As a result, it was impossible to use this `annot` to construct
bigger trees.

This issue stemmed from the fact that the implementation for pattern
quasiquotes accidentally introduced a variable called `inferredPt`
that shadowed a variable called `inferredPt` calculated in the outer
scope. As a result, the old (correct) value for pt was overwritten.