refactor handling of repeated arguments, fix #910
In scalac, handling of repeated arguments is all over the place.
Some checks are done in the parser, some checks are done in the typer,
and the rules are pretty crazy, especially given that scalac doesn't
have a dedicated node for repeated arguments.

Scalameta has a dedicated node for repeated arguments, and as a result
we had to introduce additional checks to make sure that our parser ASTs
make sense. Unfortunately, it turns out that our checks were too strict.

In this commit, we make sure that the checks have a reasonable degree
of strictness. This required a surprising amount of work.