spin off semanticdb into a separate workspace
Recently, we've been encountering more and more mismatches between the original
vision of semantic database and the reality of how semanticdb is used.

Initially, I envisioned semanticdb as an implementation detail for Scalameta
semantic API, i.e. an enabler for new-style def macros. As a result, it was
tightly coupled with the rest of Scalameta, depending on almost every module
and using a lot of our internal infrastructure.

As time went on, we figured that semanticdb is incredibly useful on its own.
First, there was scalafix, then we came up with messagehost, then it was
metadoc, and finally metaide. All those use cases chipped away at semanticdb's
ties with Scalameta, until today I finally got fed up and factored out
semanticdb codebase from Scalameta. I would've moved the resulting workspace
into a separate github organization, but the logistical hassle isn't worth it
until we have a buy-in from Olaf.

Summary of changes:
  * scalameta/io and scalameta/inputs became starmeta/io and starmeta/inputs.
    They no longer use any macro annotations and are ready to be crosscompiled
    to the variety of Scala 2.x versions as well as Dotty
    (this fun enterprise is left for future work).
  * For backward compatibility reasons, scalameta/io and scalameta/inputs
    still exist, but they are just shells of their former selves, only defining
    aliases to starmeta/io and starmeta/inputs.
  * scalameta/semantic became starmeta/semanticdb, which only exposes a
    protobuf schema and the accompanying Scala bindings. The functionality
    that doesn't fit into this scheme - i.e. parsing Scala sources and
    providing extension methods for Scalameta trees - has been dropped.
  * As a result, starmeta/semanticdb became language-agnostic.
  * Again, for backward compatibility reasons, I've reintroduced
    scalameta/semantic which simply redirects to starmeta/semanticdb.
    If we find the deleted functionality useful, we will be able to
    reintroduce it later as part of scalameta/semantic.
  * scalahost and its submodules remained untouched. Since scalahost uses
    Scalameta to parse and process the sources, it is a better fit for
    the scalameta workspace because of logistical reasons.