credit Twitter for sponsoring my participation in Scalameta
This is long overdue, but I'm glad to finally follow this through.
It is very satisfying to see the license notice carry Twitter's name.

After I joined Twitter in November 2016, I've been able to work
on Scalameta and related projects as my full-time day job. This is a very
rare opportunity, and I am truly grateful to the company for supporting
Scalameta both morally and financially.