Hack to fully qualify magic names.
The scalac pretty printer shortens the names for several symbols by
stripping out the prefix. This causes bugs like
https://github.com/scalacenter/scalafix/issues/254

This commit adds a special case in DenotationOps to add the prefix when
the printed name contains no `.`. I spent a while trying to fix this at
the root by extending the pretty printer, but gave up eventually in
favor of this hack.