Make tests succeed on Windows
Mostly newline funny business. Multiline strings are as-is and don't
aquire different line endings on Windows.

It also seems like the platform-specific line endings in
Position.format aren't actually being expanded to platform specific line
endings.

That shouldn't be causing any actual harm, so the tests just plaster
over it in an "whichever way makes it succeed" kind of way - (almost?)
all Windows tooling doesn't care much about whether newlines are \r\n or
\n

Other than that, "\foo\bar" is not an absolute path on Windows

Leaves two tests failing, of which I'm not actually sure the failures
are spurious.