Remove org.langmeta.
sbt-sonatype does not support closing staging repositories for multiple
organizations, it only picks the `organization` defined for the root
project. We'll have to figure out later how to make the release step
work for both org.langmeta and org.scalameta, but I can't let this
block a scalameta v2 release at this moment.