Make Denotation.members List[Signature] instead of Option[List[_]]
The original idea was that

- None indicated members were not stored in the semanticdb
- Some(Nil) indicated members were stored in the semanticdb but this
  denotation had no members.

However, after some thinking how this would affect downstream users I
realized it's better to provide that bit of information (members are
persisted/not persisted) in a different place like
`Document.settings.members = all`. By using Option[_], we open a can of
worms related to validating consistency, what happens for example if
one Denotation.member is Some(Nil) and another one in the same file is
None? Additionally, in protobuf, it's possible to distinguish None from
Some(Nil).