Fix message reporting to avoid memory overhead.
This commit changes the persistence of deprecation messages so that they
- are appended to existing .semanticdb files
- are merged with the original s.Document on Database.load
The implementation is quite hacky, since we have to keep track of how
many messages have been persisted to to avoid reporting the same
messages twice (once in typer and once in jvm). We use a mutable map
inside ReporterOps where we keep track of this state.