Temporarily reverts the upgrade to sbt-crossproject
Looks like there is a minor incompatibility between sbt-scalajs
and sbt-crossproject, which needs non-trivial effort to get fixed:
https://github.com/scalameta/scalameta/pull/1228#issuecomment-359680568.

While we can certainly work around this incompatibility, I suggest
that we don't get into the business of prematurely crossing bridges.

Since we're not ready for a wholesale upgrade to Scala Native,
we will be just fine with sbt-scalajs for the time being.