Crosscompile semanticdb3 and metap to a prerelease build of Scala Native
Unsatisfied by the fact that we have to depend on a temporary build
of Scala Native, I bit the bullet and published a permanent prerelease
build under my personal organization on Sonatype.

This meant that I had to rebuild the entire dependency chain
of semanticdb3 using this prerelease build and then publish them
- again under my personal organization.

All in all, this was a fun exercise.