Fix 1389
This allows the highjacked analyzer to use the old analyzer's Macro
class loader. I believe this fix is dirty and that the right thing to
do would be for the highjacked analyser to decorate the old one with
additional behaviour, but it seems like a tremendous task considering
the size of the interface.

The reason for this fix is to avoid the semanticdb plugin to temper with
the behaviour of ammonite.

See https://github.com/scalameta/scalameta/issues/1389