Remove SymbolInformation.location
Ever since its introduction more than a month ago, this field
has been unimplemented and unused. I think it's time to retire it
for the time being.

We may end very well up needing SymbolInformation.location in the future,
primarily because of our LSP work, but let's not get ahead of ourselves.