Make Symbol.Multi.syntax safe.
Exception throwing toString/syntax makes it difficult to even debug why
you have a Symbol.Multi in the first place.  Given that we use
Symbol.Multi in both semanticdb-scalac and in scalafix and we don't see
a path forward how to deprecate it, I think it's fair to make toString
safe to run.  The parser fails on multi symbols, so they still won't
sneak into applications.