Use `.par.foreach` and introduce metacp --par option.
This commit replaces the usage of `.par.flatMap` with `.par.foreach` as
an attempt to prevent a difficult-to-reproduce deadlock problem in
parallel collections.  The parallel behavior is disabled by default and
can be enabled with the --par option.

Using --par seems to yield ~40-50% faster performance in
MetacpCrashSuite on my Macbook. It takes ~36s to run with --par and ~57s
without --par.

The theory for why `.par.foreach` is better than `.par.flatMap` is based
on the experience in scalafmt/scalafix where I've used that same pattern
for over two years and have not seen reports of deadlock behavior.
I have never used `.par.flatMap` in my applications.