Merge langmeta back into scalameta
Separating langmeta from scalameta was a mistake, since our use cases
ended up not requiring this level of generality, but the abstraction
overhead (aliases, education, extra artifacts, etc) was real.

Now that we're going for a major release because of the recent removal
of the high-level semantic API, we might as well fix this mistake.