Cross-build for Scala Native.
Most of the workarounds are copy-pasted from
https://github.com/scalameta/scalameta/pull/1243
I validated that the workarounds are necessary by first trying to run
the build without them.

The commented out tests show that there is some work left to be done to
bridge the gap between JS/JVM/Native. However, given that over 99% of
the tests pass should be a big enough milestone to justify merging and
leave the remaining issues for followup PRs.