Add materialized types to symbol occurrences
New mode for symbol config is added that when enabled fills materializedType filed in each symbol occurence that is reference (there is no point in filling it on definition side since type will be the same as one stored for symbol)

In order to obtain materialized type I sometimes need to access outer tree (e.g. for tpe for `empty` in `List.empty[Int]` is generic and materialized type is stored in outer TypeApply tree). That is why I needed to introduce processed trees stack.