Apply typer diffs from 2.12.4 to 2.12.6.
This diff is produced by manually reviewing each individual change in
the following diff from scala/scala

https://gist.github.com/olafurpg/864d497882689e0b8d0a0d25f4840c08

The best way to review this diff is to go over the changes in this
commit and that gist one line at a time. Note that many changes in the
diff do not need handling in SemanticdbAnalyzer since those changes are
for methods that are not overridden by the hijacked analyzer.