Rename -P:semanticdb:failures to -P:semanticdb:crashes
Failures is not specific enough and is also ambiguous since
we are sometimes saving failures (aka errors) into the "Diagnostics"
section of SemanticDB payloads.

This ambiguity wasn't a problem when we added this option
(at the very dawn of the SemanticDB compiler plugin), but we've grown
significantly in the last 1.5 years, so it's time for some polish.