Change -P:semanticdb:mode to -P:semanticdb:text
Continuing with plugin options cleanup, I've changed `mode` to `text`.
The reasons for this change are two-fold.

First, `mode` made sense when we had just one mode in the plugin.
Now we have at least half a dozen.

Second, `-P:semanticdb:mode:disabled` is cute, but we now have filters
that can be used towards the same end. Moreover, if someone wants to
disable the plugin via compiler flags, they are free to not add the plugin
to compiler flags in the first place.