No longer emit occurrences for selections from structural types
This one is very non-obvious, so let's get into details.
Let's take an example and consider what should be the targeted
occurrence in the following code:

  val c = new C { val x: Int = ??? }
  c.<<x>>

Scalameta 3.7.4 thinks that we should emit an occurrence with
the symbol equal to the symbol of `val x: Int = ???`.
This is neat, but can this really work?

We have faced one of the complications in #1426 - if we try to
naively create global symbols for members of refinement classes,
we run into naming conflicts. It is probably possible to resolve
these conflicts with a smarter naming scheme, but I have found that
this is all moot.

I believe that trying to correlate symbols of these two occurrences
is unsound. Let's modify our example a little bit:

  type T = { val x: Int }
  val c: T = new C { val x: Int = ??? }
  c.<<x>>

Which symbol should the targeted occurrence carry? Should it still
be the symbol of `val x: Int = ???`? That doesn't make any sense.
Well, maybe let's try the symbol of `val x: Int`?

That's still not good enough though, as illustrated by the
following extension to our running example:

  type T = { val x: Int }
  val c: T = new C { val x: Int = ??? }
  def id[T](x: T): T = x
  id(c).<<x>>

In this code, there's literally no definition that could have the
same symbol as <<x>>.

If we get really insistent, we could try to correlate <<x>> with an
inferred type argument that we smuggle via a synthetic. However that's
a huge stretch, and an unsound one at that, because the "Synthetics"
section is totally optional.

Because of all these difficulties, I decided to stop emitting
occurrences for selections from structural types. If that becomes
a problem, we can always extend SymbolOccurrence to carry a type or
at least an info (or we maybe we can create synthetic symbols just
for occurrences like that), but I don't want to overcomplicate things
right now.

Thinking about this in a more general fashion, it would seem that
our symbol-based semantic model has some rough edges when it encounters
symbol-carrying types (Scalac has similar difficulties, too).

Unfortunately, Scala has plenty of symbol-carrying types:
1) structural types, 2) existential types, 3) poly types,
4) class info types, 5) method types, and 6) type types.
Luckily for us, only the first two are value types, and existential
types aren't particularly fond of showing the world their skolems
(all my attempts to expose naked skolems ended with Scalac typer
getting rid of them via existentialTransform).