Introduce LinkMode in metac
In subsequent commits, we want to change persistence of StructuralType,
ExistentialType and UniversalType to hardlink their children (i.e.
declarations, type parameters and parameters), as well as children
of their children, etc.

To prepare for that, we introduce LinkMode and add LinkMode parameters
to g.Symbol.toSymbolInformation, g.Type.toSemantic and some others.

No behavior is changed in this commit, apart from order of some gensyms.
That is left to subsequent commits to simplify code review.