Fix scalacp references to static inner Java classes
This commit fixes a bug that resulted in scalacp producing references
external static Java inner classes as if they were Scala object symbols.
For example, previously scalacp produced the symbol `java.util.Map.Entry#`
while javacp produced the expected symbol `java.util.Map#Entry#`.

This issue turned out to be tricky to solve because Scala signatures
don't distinguish between references to external static Java inner
classes and Scala inner classes inside of objects. The solution in this
commit is to require the full dependency classpath in metacp and query
the classfiles whether they are Java or Scala defined.

The overhead of scanning the full classpath is somewhat costly but
manageable, for a classpath of ~200mb it takes around 200ms on my
laptop. I expected the overhead to be more significant.

While fixing this issue, I had to implement missing functionality to
handle classpaths:

- `PathIO.{basename,dirname}` to extract directory and base name of string paths
- `ClasspathEntry` as an abstraction over a single package, file or zip
  entry in the classpath.
- `ClasspathLookup` to scan a full classpath and resolve paths to
  `ClasspathEntry`.  I started by using `java.net.URLClassLoader` but it
  it not correctly handle jar files that are missing directory entries
  (which includes the JDK jars).  ClasspathLookup can handle classpaths
  entries that are directories, regular jars and manifest jars.

The default metacp behavior for missing symbols is to fail fast with a
helpful error message. However, some classpaths can be incomplete for
several reasons. For example, some libraries embed shaded dependencies
in which case the Scala signatures will point to non-existing
classfiles.  To be able to recover from missing symbols, metacp allows
users to provide the flags `--assume-java <package>` and
`--assume-scala <package>` to tell metacp to assume symbols in those
packages are Scala/Java defined in case they are missing from the
classpath. This option was required to keep ConvertSuite green, since
both akka-stream and spark-sql referenced symbols that were missing in
the classpath.