Don't silently assume k.OBJECT for missing dependency symbols.
Fixes #1669. This commit changes the scalacp detection of packages vs.
objects by failing fast if the symbol does not exist in the classpath
instead of silently falling back to assume that the symbol is an object.

This fix highlighted that --assume-scala and --assume-java are unsafe
after we started using slashes / for package symbols. Even if we know a
symbol is Java defined we need to know which symbols are actual package
symbols.  The --assume options have been removed and ConvertSuite has
been updated to declare all dependencies of the tested libraries.

This change means that we can't process akka-streams since it has APIs
that depend on JDK 9.