Create new metai command-line tool
The tool can be used like this
```
metai [<classpath> ...]
```
and it will write in-place the file `META-INF/semanticdb.semanticidx`
with a lookup index from toplevel symbols to the relative URIs where
those files are contained.

This commit

- fixes `RelativePath.toURI` to return a proper relative URI
- introduces `Classpath.foreach(AbsolutePath => Unit)` helper
  to traverse a classpath handling jar files and manifests.

The tests pass except that metai.expect does not contain entries from
manifest.jar.