Welcome Max Ovsiankin to the team!
Max made a strategically important contribution to the Scalameta
vision by implementing a Javac compiler plugin that produces SemanticDB
payloads.

This is a critical advance of our vision to make SemanticDB a polyglot
format for semantic information, and we'd like to officially recognize
that by making Max part of the Scalameta team.