Reintegrate ScalaPB into the build.
This commit upgrades too ScalaPB v0.8.0-RC1 making the protobuf
generation automatic again. Previously, we manually ran the
codegeneration to take advantage of the new sealed oneof encoding
before it was released.

This change means that "Goto definition" in IntelliJ no longer shows the
generated sources since sbt does not include `unmanagedSources` in the
-sources.jar. I'm sure it's possible to configure sbt to include the
generated sources, but let's leave it for a separate PR.