Update the spec for sneaky changes in the local symbol format
Back then, we sneakily changes the local symbol format to also include pluses.
That was done to simplify setter symbol synthesis for local private[this] vars
(e.g. if the corresponding getter symbol is `local123`, the setter symbol
is `local123+1` and its parameter is `local123+2`).

This commit reflects this change in both the SemanticDB specification
and the reference implementation of the symbol format in Scala.scala.