Implement utility method to print single symbol info
While working on scalafix.v1 it was suggested to re-use metap for
pretty-printing. Previously, metap required a TextDocument to eagerly
build a mini-symtab to print symbols. After this commit, it's possible to
use a lazy GlobalSymbolTable instead. Additionally, the access for
`InfoNotes` and `InfoPrinter` is now public so that it's possible
implement custom formatting like skipping the symbol. These are internal
APIs anyways so we promise no bin or source compat.