Disable typer instrumentation by default
Instrumentation of typer has caused problems in the past

- difficult to upgrade compiler versions
- incompatibilities with other compiler plugins

While testing metals in the spotify/scio and typelevel/cats repos, I
encountered infinite loops during scalamacros/paradise expansion caused
by typer instrumentation.  This problem was the last drop in the bucket
in my opinion and enough justification to disable it by default.

Disabling instrumentation will result in regressions for several cases.
The best approach to recover from those regression is to contribute
upstream to scala/scala the attachments that the current instrumenter
inserts. We can already take advantage of the existing
`OriginalTreeAttachment` that records constant folding.