Goodbye, Scalameta!
I'm very proud of what we've achieved in Scalameta.

What started as a research macro system for Scala, ended up as a foundation
for a family of new-style developer tools, with more than 10000 commits
and 200 contributors. Even more, one of our technologies - SemanticDB -
has become a metadata format for Rsc, a brand new Scala compiler.
I couldn't have wished for more!!

Olaf, the stage is yours.