Emit SymbolInformation for anonymous classes.
Previously, the symbols section didn't contain information about
anonymous classes preventing us from implementing "rename symbol" in
Metals. For example, when renaming a method that is overridden in an
anonymous class then "rename symbol" wouldn't find that override.

Now, we emit a `SymbolInformation` for anonymous classes so that rename
symbol knowns all the overridden implementations of a method.