Turn `allowNumericLiteralUnderscoreSeparators` into a var for bincompat
Previously, the `allowNumericLiteralUnderscoreSeparators` setting was a
regular field in the `Dialect` case class like all the other dialect
settings. However, since Scalameta v4 we started taking binary
compatibility seriously and this means that adding a new field to the
`Dialect` case class would require us to publish Scalameta v5 next
because the signature of the `Dialect` constructor and `copy()` method
change with every added/removed field.

To avoid breaking binary compatiblity, this commit turns the new setting
into a var field that can be configured via a builder method
`withAllowNumericLiteralUnderscoreSeparators(true): Dialect`. The var field is
private and can only be mutated on fresh `Dialect` instances.