Get rid of global token cache, fixes #1068
Previously, Scalameta kept a global cache for the tokens of all inputs.
Now, the global cache has been replaced with a cache that's local to the
input itself.

The new solution was not possible before #1863 since the `Input` type
was kept in a separate module from `Tokens`. Now the two types live in
the same module so it's possible to make the token cache local to
inputs.

I verified that this change does not regress the performance of
Scalafmt.