Move docs from tutorial repo into main repo
Previously, we maintained the scalameta docs in a separate repo. This
setup relied on a fairly complex setup downloading markdown files from
the scalameta repo. The documentation in the Scalameta repo was also
lying in several different directories

* `notes/`
* `semanticdb/semanticdb3/`

Now, all the docs for scalameta.org lie in the scalameta/scalameta repo
under the `docs/` repository, using the same convention as we have in
other repositories.