Fix symbols of nullary method eta-expansions
When the Scala parser encounters an underscore while parsing expressions
it discards that token's position and creates a `MethodValue` of the
target method:

    case USCORE =>
      atPos(t.pos.start, in.skipToken()) { MethodValue(stripParens(t)) }

Then the Scala typechecker creates a `Function0` out of that, defining
its position as the position of the underlying method value:

    val pos = methodValue.pos
    // must create it here to change owner (normally done by typed's typedFunction)
    val funSym = context.owner.newAnonymousFunctionValue(pos)
    new ChangeOwnerTraverser(context.owner, funSym) traverse methodValue

    val result = typed(Function(List(), methodValue) setSymbol funSym setPos pos, mode, pt)

The result is that `XtensionCompilationUnitDocument#toTextDocument`,
which uses positions, incorrectly matches the `Function0` (Global) tree
(e.g. `(() => this.prop)`) to the underlying method expression (`prop`)
Scalameta tree, and then converts that anonymous class (Global) symbol
into a local Semanticdb symbol.

So, instead of that, we skip the other function when it has no params
(by flipping from `exists` to `forall`) so now the Scalameta tree `prop`
is correctly associated to the method's symbol.