Make it possible to evolve `Dialect` in a binary compatible way.
Previously, we used a hacky solution based on private vars to introduce
new fields to the `Dialect` class in a binary compatible way. This hacky
solution leaked to users when calling the auto-generated `copy()` method
because it would not copy over the private mutable fields.

Now, the `Dialect` class no longer uses the `@data` annotation and
instead this commit manually implements the macro expansion. This allows
us to introduce new constructor parameters without breaking the binary
signature of the previously auto-generated `copy()` method. This change
also fixes the situation so that users no longer need to worry about the
order in which they customize fields in `Dialect.copy(...).withX(...)`.